Move test/utils.go to utils_test.go
